fix(v0.14): flow Long-parent dtype + dotted-node-name ModuleDict bugs (#200)
* fix(v0.14): flow Long-parent dtype + dotted-node-name ModuleDict bugs

Two minimal fixes recovering NBN coverage on bnlearn networks that
previously crashed (surfaced by the bnlearn_complete investigation).

Bug 1 (flow, Long discrete parents): a flow node with a discrete
parent crashed with 'mat1 and mat2 must have the same dtype' because
the inference entry points passed a Long context into the float
conditioner. Cast the parent context to float in forward / log_prob /
sample (mirrors fit_local's cast and MDN's implicit promotion). Fixes
nbn-flow-* on clg healthcare / sangiovese.

Bug 2 (dotted node names): nn.ModuleDict forbids '.' in keys, so every
NBN mechanism crashed on networks with dotted node names (mehra's
pm2.5, magic-niab's YR.GLASS). Add _AliasModuleDict, an nn.ModuleDict
that stores a dot-free alias ('.' -> '__DOT__') and reverses it on read
via the mapping protocol (getitem/contains/iter/keys/items), so every
call site keeps using the original node name unchanged — a single-point
fix at construction, no read sites touched. The transform is pure and
reversible (no side map to desync across deepcopy / load_state_dict).

Tests: flow log_prob/forward/sample with a Long parent; _AliasModuleDict
round-trip + end-to-end fit+query on a dotted-name net under both a
plain-table and a neural mechanism. Verified on real bnlearn networks:
flow/healthcare and mdn-lw/mehra (pm2.5) both fit + query cleanly.


* fix(lint): clear ruff E741 + SIM118 so PR-A CI is green

- E741 (PR #199): rename ambiguous `l` -> `lbl` in the duplicate-labels
  assertion (tests/unit/v013/test_paper_figures.py).
- SIM118 (PR-A): _AliasModuleDict.keys() iterated super().keys(); use
  list(self) instead (__iter__ already de-aliases). [PR-A]


---------

+++ nbn/mechanisms/normalizing_flow.py
@@ -147,7 +147,10 @@ class NormalizingFlowMechanism(Mechanism):
         assert self._flow is not None, "Call fit_local before forward()."
         if self._d_pa == 0 or parents is None:
             return _FlowDistribution(self._flow(None))
-        return _FlowDistribution(self._flow(ensure_2d(parents)))
+        # Cast context to float: a discrete (Long) parent would otherwise hit
+        # F.linear(Long, Float) in the conditioner. Mirrors fit_local's cast
+        # and MDN's implicit float-promotion via standardization.
+        return _FlowDistribution(self._flow(ensure_2d(parents).float()))
 
     def log_prob(self, x: torch.Tensor, parents: torch.Tensor | None) -> torch.Tensor:
         assert self._flow is not None
@@ -161,6 +164,7 @@ class NormalizingFlowMechanism(Mechanism):
         if self._d_pa == 0 or parents is None:
             ctx = None
         else:
+            parents = parents.float()  # discrete (Long) parents -> float context
             if parents.dim() == 2:
                 parents = parents.unsqueeze(1).expand(-1, s, -1)
             flat, _, _ = flatten_samples(parents)
@@ -177,7 +181,7 @@ class NormalizingFlowMechanism(Mechanism):
             with torch.no_grad():
                 samp = self._flow(None).sample((b * n,))  # [B*n, D_x]
         else:
-            ctx = ensure_2d(parents).unsqueeze(1).expand(-1, n, -1)
+            ctx = ensure_2d(parents).float().unsqueeze(1).expand(-1, n, -1)
             flat, _, _ = flatten_samples(ctx)
             with torch.no_grad():
                 samp = self._flow(flat).sample()  # [B*n, D_x]
